Refactor shortcut code files
